+++ lib/esm/transformations/SuperResolution.d.ts
@@ -1,6 +1,8 @@
 export function upscale(config?: {
     type: string;
     enhanceFace: boolean;
+    model: string;
+    enhanceQuality: boolean;
 }): Transformation;
 declare namespace _default {
     export { upscale };